requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
einops==0.6.0
fonttools==4.38.0
ftfy==6.1.1
idna==3.4
imageio==2.23.0
kiwisolver==1.4.4
matplotlib==3.6.2
networkx==2.8.8
numpy
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.6.0.66
packaging==22.0
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets
regex==2022.10.31
requests==2.28.1
scikit-image
scipy
six==1.16.0
tifffile==2022.10.10
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
